+++ lib/hexapdf/layout/inline_box.rb
@@ -47,9 +47,10 @@ module HexaPDF
     # beforehand! This means the box *must* have at least its width set. The height may either also
     # be set or determined during fitting.
     #
-    # Fitting of the wrapped box is done immediately after creating a InlineBox instance. For this,
-    # a frame is used that has the width of the wrapped box and its height, or if not set, a
-    # practically infinite height. In the latter case the height *must* be set during fitting.
+    # Fitting of the wrapped box via #fit_wrapped_box needs to be done before accessing any other
+    # method that uses the wrapped box. For fitting, a frame is used that has the width of the
+    # wrapped box and its height, or if not set, a practically infinite height. In the latter case
+    # the height *must* be set during fitting.
     class InlineBox
 
       # Creates an InlineBox that wraps a basic Box. All arguments (except +valign+) and the block
@@ -76,12 +77,6 @@ module HexaPDF
         raise HexaPDF::Error, "Width of box not set" if box.width == 0
         @box = box
         @valign = valign
-        @fit_result = Frame.new(0, 0, box.width, box.height == 0 ? 100_000 : box.height).fit(box)
-        if !@fit_result.success?
-          raise HexaPDF::Error, "Box for inline use could not be fit"
-        elsif box.height > 99_000
-          raise HexaPDF::Error, "Box for inline use has no valid height set after fitting"
-        end
       end
 
       # Returns the style of the wrapped box.
@@ -129,6 +124,17 @@ module HexaPDF
       # The maximum y-coordinate which is equivalent to the height of the inline box.
       def y_max
         height
+      end
+
+      # Fits the wrapped box, using the given context (see Frame#context).
+      def fit_wrapped_box(context)
+        @fit_result = Frame.new(0, 0, box.width, box.height == 0 ? 100_000 : box.height,
+                                context: context).fit(box)
+        if !@fit_result.success?
+          raise HexaPDF::Error, "Box for inline use could not be fit"
+        elsif box.height > 99_000
+          raise HexaPDF::Error, "Box for inline use has no valid height set after fitting"
+        end
       end
 
     end