+++ lib/hexapdf/layout/text_box.rb
@@ -82,13 +82,13 @@ module HexaPDF
         @width = @height = 0
         @result = if style.position == :flow
                     @tl.fit(@items, frame.width_specification, frame.shape.bbox.height,
-                            apply_first_text_indent: !split_box?)
+                            apply_first_text_indent: !split_box?, frame: frame)
                   else
                     @width = reserved_width
                     @height = reserved_height
                     width = (@initial_width > 0 ? @initial_width : available_width) - @width
                     height = (@initial_height > 0 ? @initial_height : available_height) - @height
-                    @tl.fit(@items, width, height, apply_first_text_indent: !split_box?)
+                    @tl.fit(@items, width, height, apply_first_text_indent: !split_box?, frame: frame)
                   end
         @width += if @initial_width > 0 || style.align == :center || style.align == :right
                     width